+++ pyxel/pipelines/processor.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -18,7 +18,6 @@ import numpy as np
 from pyxel import __version__
 from pyxel.evaluator import eval_entry
 from pyxel.pipelines import DetectionPipeline, ModelGroup
-from pyxel.state import get_obj_att
 from pyxel.util import get_size
 
 if TYPE_CHECKING:
@@ -85,6 +84,74 @@ def result_keys(result_type: ResultId) -> Sequence[ResultId]:
     return [ResultId(result_type)]
 
 
+# TODO: Refactor this function (e.g. include it in 'Processor')
+def _get_obj_att(
+    obj: Any, key: str, obj_type: Optional[type] = None
+) -> tuple[Any, str]:
+    """Retrieve an object associated with a specified key.
+
+    The function is versatile and can be applied to dictionaries, lists,
+    or user-defined objects, such as configuration models.
+
+    Parameters
+    ----------
+    obj : Any
+        The target object from which to extract the desired attribute.
+    key : str
+        A string representing the attribute path within the object.
+        Nested attributes are separated by dots.
+    obj_type
+        An optional parameter specifying the expected type of the retrieved object.
+        If provided, the function ensures that the final object matches this type.
+
+    Returns
+    -------
+    A tuple containing two elements
+
+    1. The object associated with the specified key.
+    2. The name of the attribute, extracted from the key.
+
+    Examples
+    --------
+    >>> obj = {"processor": {"pipeline": {"models": [1, 2, 3]}}}
+    >>> get_obj_att(obj, "processor.pipeline.models")
+    ({'models': [1, 2, 3]}, 'models')
+    """
+    *body, tail = key.split(".")
+    for part in body:
+        try:
+            if isinstance(obj, dict):
+                obj = obj[part]
+            elif isinstance(obj, list):
+                try:
+                    index = int(part)
+                    obj = obj[index]
+                except ValueError:
+                    for _, obj_i in enumerate(obj):
+                        if hasattr(obj_i, part):
+                            obj = getattr(obj_i, part)
+                            break
+                        elif obj_i.__class__.__name__ == part:
+                            if hasattr(obj_i, tail):
+                                obj = obj_i
+                                break
+            elif hasattr(obj, part):
+                obj = getattr(obj, part)
+            else:
+                raise NotImplementedError(
+                    f"obj={obj!r}, key={key!r}, obj_type={obj_type!r}, part={part!r}"
+                )
+
+            if obj_type and isinstance(obj, obj_type):
+                return obj, tail
+
+        except AttributeError:
+            # logging.error('Cannot find attribute %r in key %r', part, key)
+            obj = None
+            break
+    return obj, tail
+
+
 # TODO: Is this class needed ?
 class Processor:
     """TBW."""
@@ -120,7 +187,7 @@ class Processor:
         True
         """
         found = False
-        obj, att = get_obj_att(self, key)
+        obj, att = _get_obj_att(self, key)
         if isinstance(obj, dict) and att in obj:
             found = True
         elif hasattr(obj, att):
@@ -148,7 +215,7 @@ class Processor:
         return result
 
     # TODO: Could it be renamed '__setitem__' ?
-    def set(  # noqa: A003
+    def set(
         self,
         key: str,
         value: Union[str, Number, np.ndarray, Sequence[Union[str, Number, np.ndarray]]],
@@ -188,7 +255,7 @@ class Processor:
         else:
             new_value = value
 
-        obj, att = get_obj_att(self, key)
+        obj, att = _get_obj_att(self, key)
 
         if isinstance(obj, dict) and att in obj:
             obj[att] = new_value
@@ -196,12 +263,12 @@ class Processor:
             setattr(obj, att, new_value)
 
     # TODO: Create a method `DetectionPipeline.run`
-    def run_pipeline(self, with_intermediate_steps: bool = False) -> None:
+    def run_pipeline(self, debug: bool) -> None:
         """Run a pipeline with all its models in the right order.
 
         Parameters
         ----------
-        with_intermediate_steps : bool
+        debug : bool
 
         Notes
         -----
@@ -218,7 +285,7 @@ class Processor:
             self._log.info("Processing group: %r", group_name)
             models_grp.run(
                 detector=self.detector,
-                with_intermediate_steps=with_intermediate_steps,
+                debug=debug,
             )
 
     # TODO: Refactor '.result'. See #524. Deprecate this method ?